# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SEAPEDIA API",
    "version": "1.0.0",
    "description": "Multi-role e-commerce platform API. Supports Admin, Seller, Buyer, and Driver workflows. Protected endpoints require a Bearer JWT token in the Authorization header."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/orders/checkout/validate-voucher": {
      "post": {
        "tags": [
          "Orders - Buyer"
        ],
        "summary": "Validate voucher code",
        "description": "Checks if code is valid for current cart's store. Returns discount amount.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code"
                ],
                "properties": {
                  "code": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Valid code with discount info"
          },
          "400": {
            "description": "Invalid or expired"
          }
        }
      }
    },
    "/orders/checkout": {
      "post": {
        "tags": [
          "Orders - Buyer"
        ],
        "summary": "Checkout",
        "description": "Converts cart to order. Calculates: subtotal, discount, tax (PPN 12% on subtotal-discount), delivery fee, final total. Methods: INSTANT, NEXT_DAY, REGULAR.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "addressId",
                  "deliveryMethod"
                ],
                "properties": {
                  "addressId": {
                    "type": "string"
                  },
                  "deliveryMethod": {
                    "type": "string",
                    "enum": [
                      "INSTANT",
                      "NEXT_DAY",
                      "REGULAR"
                    ]
                  },
                  "promoCode": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created with initial status SEDANG_DIKEMAS"
          },
          "400": {
            "description": "Insufficient balance or invalid cart"
          }
        }
      }
    },
...
    "/orders/{id}": {
      "get": {
        "tags": [
          "Orders - Buyer"
        ],
        "summary": "Order detail",
        "description": "Full order with items, address, status history with timestamps.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order detail"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "tags": [
          "Orders - Buyer"
        ],
        "summary": "Buyer confirms order received",
        "description": "Changes status to PESANAN_SELESAI. Releases seller payment (gross minus 5% platform commission).",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order finalized, seller wallet credited"
          }
        }
      }
    },
...
  }
}